requirements.txt
Pillow
requests
extcolors
bs4